操作補助機能：盤面選択時のナンバーハイライト機能追加
